xloader: Panda - fix ROM OTG boot path
When the OMAP 4 on Panda sees that USB OTG external power is present,
it changes the boot flow in the ROM to try to boot via that before
trying MMC1.  If it isn't satisfied, it falls through to try MMC1.

Without this patch, under those conditions X-Loader is unable to
update the DPLL as it needs to do during dram initialization, because
the EMIF never goes idle allowing the shadow update mechanism to
apply the new settings.

It therefore loops for a long period waiting to confirm the settings
applied and boot is hung (with LEDs off, so the device looks OFF).

This patch re-enables EMIF idle which was previously commented out with
"No IDLE: BUG in SDC" to allow the new DPLL settings to apply.

With the patch against http://gitorious.org/x-loader HEAD the Panda is
able to boot from either USB OTG power applied or DC Jack power OK.

Note that I saw peak current during boot of 650mA, higher than the
500mA allowed by USB.  It may go higher with other peripherals enabled
or trying to use USB host on Panda.

Signed-off-by: Andy Green <andy.green@linaro.org>