igep00x0, beagle and omap3evm: fixing bug at prcm_init
Commit 73eb46434657c8c2c9e7c9146406031fc4d124ce changed get_cpu_rev()
to be a generic function for all omap 3 based boards, to remove
duplicated code. As a side effect, the behavior changed for beagle and
omap3evm when the board is based on 37XX, returning a different value
then CPU3430_ES2, causing a bug at prcm_init.

The fix is already applied for overo, as you can see from commit
24b8b7f41a83540433024854736518876257672c, changing the index calculation
to behave correctly with all board revs. This patch only applies the
same fix for the other omap 3 based boards that could be affected.

BugLink: https://bugs.launchpad.net/ubuntu/+source/x-loader/+bug/813407

Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@linaro.org>
Acked-by: Enric Balletbo i Serra <eballetbo@gmail.com>
Signed-off-by: Anand Gadiyar <gadiyar@ti.com>